requirements.txt
# Core dependencies
asyncio
aiohttp>=3.8.0
aiofiles>=23.0.0

# AI/ML providers
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.11.0

# Utilities
pydantic>=2.0.0
tqdm>=4.65.0
click>=8.1.0
python-dotenv>=1.0.0

# Scientific computing
scipy>=1.10.0

# Testing and development
pytest>=7.4.0
pytest-asyncio>=0.21.0
